Refactor analytics modules into packages
