+++ analytics/anomaly_detection/__init__.py
@@ -1,0 +1,29 @@
+"""Anomaly detection subpackage."""
+
+from .analyzer import (
+    AnomalyDetector,
+    AnomalyDetection,
+    create_anomaly_detector,
+    EnhancedAnomalyDetector,
+)
+from .data_prep import prepare_anomaly_data
+from .statistical_detection import (
+    detect_frequency_anomalies,
+    detect_statistical_anomalies,
+    calculate_severity_from_zscore,
+)
+from .ml_inference import detect_ml_anomalies
+from .types import AnomalyAnalysis
+
+__all__ = [
+    "AnomalyDetector",
+    "AnomalyDetection",
+    "create_anomaly_detector",
+    "EnhancedAnomalyDetector",
+    "prepare_anomaly_data",
+    "detect_frequency_anomalies",
+    "detect_statistical_anomalies",
+    "calculate_severity_from_zscore",
+    "detect_ml_anomalies",
+    "AnomalyAnalysis",
+]
